refactor(handler): split conversations, skills, subagents, marketplace under 300 line limit
